Small README fixes
Hi! 👋 
I just tried out this lib. It's really easy to integrate 👍 

There are two compile errors in the Capsule Chart example, and the `dates` and `hours` in the `LineChartController` would be less surprising with a small comment.
Here's my proposal how to make it even easier to get started 😊 

Best
Jannik